feat(registry): add static registry.json catalog generated from app.json
Studio's registry/store is moving to consume a single static JSON catalog
instead of per-org Postgres rows + the MCP-binding fan-out. This adds the
first-party half of that catalog as a committed registry.json built from every
MCP's app.json, plus a generator and a CI drift check.

- scripts/lib/registry-payload.ts: extract the app.json -> registry item
  mapping into one shared module (was inline in publish-one.ts), so the
  committed registry.json stays byte-consistent with what publishing produces.
- scripts/publish-one.ts: refactor to use the shared mapping (behavior-identical).
- scripts/build-registry-json.ts: build registry.json from all app.json files
  (public items only, sorted by id, deterministic). Supports --check.
- registry.json: the generated catalog (111 public items).
- checks.yml: fail CI when registry.json is stale (run bun run build:registry).

Consumed by Studio mesh over HTTPS (commit-pinned). The community/official feed
stays in the existing Supabase-backed registry server (separate source).
